Fix Announcer VO when playing as squads

Approved-by: Carlos

+++ Assets/Src/FirstLight/Game/Utils/GameConstants.cs
@@ -69,8 +69,8 @@ namespace FirstLight.Game.Utils
 			public const float BR_MID_PHASE_SECONDS_THRESHOLD = 90f;
 			public const float BR_HIGH_PHASE_SECONDS_THRESHOLD = 180f;
 			
-			public const float DM_HIGH_PHASE_KILLS_LEFT_THRESHOLD = 3;
-			public const float BR_HIGH_PHASE_PLAYERS_LEFT_THRESHOLD = 2;
+			public const int DM_HIGH_PHASE_KILLS_LEFT_THRESHOLD = 3;
+			public const int BR_HIGH_PHASE_PLAYERS_LEFT_THRESHOLD = 2;
 			public const float HIGH_LOOP_TRANSITION_DELAY = 2f;
 			
 			public const float LOW_HP_CLUTCH_THERSHOLD_PERCENT = 0.1f;
